+++ Proyecto1/src/Persona/Cliente.java
@@ -1,36 +1,12 @@
 package Persona;
 
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.time.LocalDate;
 
 public class Cliente extends Persona {
 
-	private static final String NOMBREARCHIVO = "clientes.txt";
-
-	public Cliente(String nombre, String login, String password) {
-		super(nombre, login, password);
+	public Cliente(String nombre, String login, String password, LocalDate fechaNacimiento) {
+		super(nombre, login, password, fechaNacimiento);
 		
-	}
-
-		public void persistencia(String nombre, Persona persistirPersona){
-
-		crearArchivo(nombre);
-		guardarCliente(persistirPersona);
-
-	}
-
-		public void guardarCliente(Persona nombreAdministrador){
-
-		
-		try (BufferedWriter clienteEscrito = new BufferedWriter(new FileWriter(NOMBREARCHIVO, true))){
-			String clienteFormatoTexto = "Nombre: " + nombre + ", Login: " + login + ", Password: " + password;
-			clienteEscrito.write(clienteFormatoTexto);
-			clienteEscrito.newLine();
-
-		} catch(IOException e){
-			System.err.println("No se pudo guardar el cliente");
-		}
 	}
